# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tripzant API Documentation",
        "description": "API endpoints for Tripzant Mobile App",
        "contact": {
            "email": "support@tripzant.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Local Development Server"
        },
        {
            "url": "https://tripnstays.com/api/v1",
            "description": "Staging/Live Server"
        }
    ],
    "paths": {
...
        "/flights/book": {
            "post": {
                "tags": [
                    "Flights"
                ],
                "summary": "Book a Flight",
                "description": "Create a flight booking for the authenticated user",
                "operationId": "9608ee10c67f93487109270f4b799fea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "flight_id",
                                    "passengers"
                                ],
                                "properties": {
                                    "flight_id": {
                                        "type": "string",
                                        "example": "12345"
                                    },
                                    "passengers": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking successful"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/flights/bookings": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "My Flight Bookings",
                "description": "Get list of flight bookings for the authenticated user",
                "operationId": "018c4d9b581839d144fb35cc3d676995",
                "responses": {
                    "200": {
                        "description": "List of bookings"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/hotel/book": {
            "post": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Book a Hotel",
                "description": "Create a hotel booking for the authenticated user",
                "operationId": "4a1508448339809778bf674ec32d529a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "hotel_code",
                                    "room_type",
                                    "check_in",
                                    "check_out",
                                    "guests",
                                    "rate_key",
                                    "holder_name",
                                    "holder_surname",
                                    "email"
                                ],
                                "properties": {
                                    "hotel_code": {
                                        "type": "string"
                                    },
                                    "room_type": {
                                        "type": "string"
                                    },
                                    "check_in": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "check_out": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "guests": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    },
                                    "rate_key": {
                                        "type": "string"
                                    },
                                    "holder_name": {
                                        "type": "string"
                                    },
                                    "holder_surname": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking successful"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/hotels/bookings": {
            "get": {
                "tags": [
                    "Hotels"
                ],
                "summary": "My Hotel Bookings",
                "description": "Get list of hotel bookings for the authenticated user",
                "operationId": "64b7e9ea2d3c7722a5abdc832680932b",
                "responses": {
                    "200": {
                        "description": "List of hotel bookings"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}